requirements.txt
timm
torchvision
dghs-imgutils>=0.17.0
torch>=2.6.0
pillow
numpy
scipy
matplotlib
huggingface_hub
hfutils>=0.11
hbutils
datasets>=3.5.0
pandas
tqdm
safetensors
tensorboard
di-toolkit
accelerate>=1.4.0
wandb
onnx
onnxoptimizer
onnxsim
onnxruntime
thop
tabulate
calflops
transformers
